requirements.txt
torch~=2.2.2
matplotlib~=3.8.4
omegaconf~=2.3.0
mlflow~=2.11.3
torchinfo~=1.8.0
numpy~=1.26.4
tqdm~=4.66.2
scipy~=1.13.0
pandas~=2.2.1
pyro-ppl~=1.9.0
hydra-core~=1.3.2
lightning~=2.2.2
sdeint~=0.3.0
torchmetrics~=1.3.2